Rename JSLint to JSLintUtils
