+++ src/debug/DebugCommandHandlers.js
@@ -10,13 +10,13 @@ define(function (require, exports, module) {
     
     var Commands                = require("command/Commands"),
         CommandManager          = require("command/CommandManager"),
-        JSLint                  = require("language/JSLint"),
+        JSLintUtils             = require("language/JSLintUtils"),
         PerfUtils               = require("utils/PerfUtils");
     
     function _handleEnableJSLint() {
-        JSLint.setEnabled(!JSLint.getEnabled());
-        JSLint.run();
-        $("#jslint-enabled-checkbox").css("display", JSLint.getEnabled() ? "" : "none");
+        JSLintUtils.setEnabled(!JSLintUtils.getEnabled());
+        JSLintUtils.run();
+        $("#jslint-enabled-checkbox").css("display", JSLintUtils.getEnabled() ? "" : "none");
     }
     
     // Implements the 'Run Tests' menu to bring up the Jasmine unit test window
