+++ src/language/JSLintUtils.js
